requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
yt-dlp
xmltodict
feedparser
google-generativeai
requests
dnspython
pandas
google-adk
mangum
gunicorn
google-generativeai
requests
dnspython
pandas
google-adk
